\name{RegWhy-package}
\alias{RegWhy-package}
\alias{RegWhy}
\docType{package}
\title{package
\packageTitle{RegWhy}
}
\description{
\packageDescription{RegWhy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RegWhy}
\packageIndices{RegWhy}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RegWhy}

Maintainer: \packageMaintainer{RegWhy}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
